fix warnings

+++ src/main/c/util/linenoise.c
@@ -222,14 +222,14 @@ static int enableRawMode(int fd) {
     raw = orig_termios;  /* modify the original mode */
     /* input modes: no break, no CR to NL, no parity check, no strip char,
      * no start/stop output control. */
-    raw.c_iflag &= (unsigned long)~(BRKINT | ICRNL | INPCK | ISTRIP | IXON);
+    raw.c_iflag &= (unsigned)~(BRKINT | ICRNL | INPCK | ISTRIP | IXON);
     /* output modes - disable post processing */
-    raw.c_oflag &= (unsigned long)~(OPOST);
+    raw.c_oflag &= (unsigned)~(OPOST);
     /* control modes - set 8 bit chars */
     raw.c_cflag |= (CS8);
     /* local modes - choing off, canonical off, no extended functions,
      * no signal chars (^Z,^C) */
-    raw.c_lflag &= (unsigned long)~(ECHO | ICANON | IEXTEN | ISIG);
+    raw.c_lflag &= (unsigned)~(ECHO | ICANON | IEXTEN | ISIG);
     /* control chars - set return condition: min number of bytes and timer.
      * We want read to return every single byte, without timeout. */
     raw.c_cc[VMIN] = 1; raw.c_cc[VTIME] = 0; /* 1 byte, no timer */
